Initial setup for refactoring to allow distinct clients per token
