+++ lib/teamsnap/error.rb
@@ -1,0 +1,4 @@
+module TeamSnap
+  Error = Class.new(StandardError)
+  NotFound = Class.new(TeamSnap::Error)
+end
