+++ lib/teamsnap/version.rb
@@ -1,3 +1,3 @@
 module TeamSnap
-  VERSION = "1.3.1"
+  VERSION = "2.0.0"
 end
